+++ src/solve/h_matrix.rs
@@ -1,162 +1,13 @@
+mod cluster;
+mod block;
+
 use std::{collections::HashMap, rc::Rc};
 use na::{DMatrix, DVector};
 use crate::{preprocess::mesh_data::Node, Cplx};
+use cluster::Cluster;
+use block::Block;
 
 use super::aca::ACA;
-
-/// Cluster Tree, used to partition the surface in space until
-/// they contain fewer nodes than the 'cardinality' parameter of the tree
-#[derive(Clone)]
-pub struct Cluster {
-    u_bound: [f64;3],
-    l_bound: [f64;3],
-    diameter: f64,
-    indices_contained: Vec<usize>, // node indices, not eqn indices
-    sons: Vec<Rc<Cluster>>
-}
-
-impl Cluster {
-    /// Build cluster tree from nodal data
-    /// Leaf cardinality is the approximate minimum size of clusters
-    pub fn new_from(nodes: &Vec<Node>, 
-                    indices_contained: Vec<usize>, 
-                    leaf_cardinality: usize, 
-                    eqn_map: &HashMap::<usize, usize>) -> Cluster {
-        let mut cluster = Cluster {
-            u_bound: [f64::NEG_INFINITY;3],
-            l_bound: [f64::INFINITY;3],
-            diameter: 0.0,
-            indices_contained: indices_contained,
-            sons: Vec::new()
-        };
-        cluster.process_cluster(nodes, leaf_cardinality, eqn_map);
-        return cluster;
-    }
-    /// Can be called recursively to process current cluster, split if applicable,
-    /// then process those clusters
-    fn process_cluster(&mut self, 
-                       nodes: &Vec<Node>, 
-                       leaf_cardinality: usize, 
-                       eqn_map: &HashMap::<usize, usize>) {
-        // Largely adapted from https://doi.org/10.1016/S0955-7997(02)00152-2
-        // Example 2.1
-        self.update_bounds(nodes);
-        self.update_diameter();
-        self.map_nodes_to_eqns(eqn_map);
-        if self.indices_contained.len() <= leaf_cardinality {
-            // Once the cluster is finalized, map from node indices to eqn indices
-            self.map_nodes_to_eqns(eqn_map);
-            return;
-        }
-        // Determine which direction to split the cluster
-        let mut jmax: usize = 0;
-        let mut diff = 0.0;
-        let alpha = &mut self.l_bound;
-        let beta = &mut self.u_bound;
-        for j in 0_usize..3 {
-            let c = beta[j] - alpha[j];
-            if c > diff {jmax = j; diff = c;}
-        }
-        // Split cluster at gamma in jmax direction
-        let gamma = 0.5 * (alpha[jmax] + beta[jmax]);
-        let mut indx1 = Vec::<usize>::new();
-        let mut indx2 = indx1.clone();
-        for index in &self.indices_contained {
-            if nodes[*index].coords[jmax] <= gamma {
-                indx1.push(*index);
-            }
-            else {
-                indx2.push(*index);
-            }
-        }
-        //self.indices_contained = Vec::new();
-        self.sons.push(Rc::new(Cluster::new_from(nodes, indx1, leaf_cardinality, eqn_map)));
-        self.sons.push(Rc::new(Cluster::new_from(nodes, indx2, leaf_cardinality, eqn_map)));
-    }
-    fn is_leaf(&self) -> bool { return self.sons.is_empty()}
-    fn get_diameter(&self) -> f64 { return self.diameter;}
-    fn update_bounds(&mut self, nodes: &Vec<Node>) {
-        let alpha = &mut self.l_bound;
-        let beta = &mut self.u_bound;
-        for index in &self.indices_contained {
-            for j in 0_usize..3 {
-                alpha[j] = f64::min(nodes[*index].coords[j], alpha[j]);
-                beta[j] = f64::max(nodes[*index].coords[j], beta[j]);
-            }
-        }
-    }
-    /// From example 2.2
-    fn update_diameter(&mut self) {
-        let diam = &mut self.diameter;
-        *diam = 0.0;
-        for j in 0..3 {
-            *diam += f64::powi(self.u_bound[j] - self.l_bound[j],2);
-        }
-        *diam = diam.sqrt();
-    }
-    /// Get distance between two clusters, used for building block tree
-    /// From Example 2.2
-    fn get_distance(c1: &Cluster, c2: &Cluster) -> f64 {
-        let mut dist = 0.0;
-        for j in 0..3 {
-            dist += f64::powi(f64::max(0.0, c1.l_bound[j] - c2.u_bound[j]),2);
-            dist += f64::powi(f64::max(0.0, c2.l_bound[j] - c1.u_bound[j]),2);
-        }
-        return dist.sqrt();
-    }
-    fn map_nodes_to_eqns(&mut self, eqn_map: &HashMap::<usize, usize>) {
-        for idx in &mut self.indices_contained {
-            if let Some(new_idx) = eqn_map.get(idx) {*idx = *new_idx;}
-        }
-    }
-}
-
-/// Block Tree, which is used to compare distances between clusters and determine
-/// admissibility, i.e., are they separated enough to approximate their interaction
-pub struct Block {
-    rows: Rc<Cluster>,
-    columns: Rc<Cluster>,
-    admissible: bool,
-    children: Vec<Block>
-}
-
-impl Block {
-    /// Build 'Block' of interaction between rows and columns, each represented by a cluster
-    /// eta controls the required separation for admissibility, though it is not particularly
-    /// sensitive. Recommend eta [4-10] based on http://dx.doi.org/10.3970/cmes.2009.043.149
-    pub fn new_from(rows: Rc<Cluster>, columns: Rc<Cluster>, eta: f64) -> Block {
-        let mut tree = Block {
-            rows: rows,
-            columns: columns,
-            admissible: false,
-            children: Vec::new()
-        };
-        tree.process_block(eta);
-        return tree;
-    }
-    /// Check if block can be divided (is not admissible, neither cluster is a leaf)
-    #[inline]
-    fn is_divisible(&self) -> bool {
-        !(self.admissible || self.rows.is_leaf() || self.columns.is_leaf())
-    }
-    fn process_block(&mut self, eta: f64) {
-        // update admissibility 
-        let diam1 = self.rows.get_diameter();
-        let diam2 = self.columns.get_diameter();
-        let dist12 = Cluster::get_distance(&self.rows, &self.columns);
-        // admissibiliy criterion 
-        self.admissible = f64::min(diam1, diam2) <= eta * dist12;
-        // check divisibility, if divisible, do so
-        if self.is_divisible() {
-            // divide into 4 blocks according to sons of row/son clusters
-            for rson in &self.rows.sons {
-                for cson in &self.columns.sons {
-                    self.children.push(Block::new_from(rson.clone(),cson.clone(), eta));
-                }
-            }
-        }
-    }
-}
 
 /// Can be represented in reduced format (e.g., ACA)
 struct AdmissibleBlock {
@@ -276,36 +127,36 @@ impl HMatrix {
             admissible_blocks: Vec::new(),
             inadmissible_blocks: Vec::new()
         };
-        mat.load_from(block_tree, &get_row_or_column, tolerance);
+        mat.load_from(&block_tree, &get_row_or_column, tolerance);
         mat.update_norm();
         return mat;
     }
     pub fn get_num_eqn(&self) -> usize { self.num_eqn }
     pub fn get_norm(&self) -> f64 { self.norm }
     /// Process block tree into flat vectors of admissible and inadmissible blocks
-    fn load_from<F>(&mut self, block: Block, get_row_or_column: &F, tolerance: f64)
+    fn load_from<F>(&mut self, block: &Block, get_row_or_column: &F, tolerance: f64)
     where F: Fn(Vec<usize>, Vec<usize>) -> Vec::<Cplx> {
-        if block.admissible {
+        if block.is_admissible() {
             self.admissible_blocks.push(
                 AdmissibleBlock::new(
-                    block.rows.indices_contained.clone(),
-                    block.columns.indices_contained.clone(),
+                    block.get_row_indices().clone(),
+                    block.get_column_indices().clone(),
                     get_row_or_column,
                     tolerance
                 )
             );
         }
-        else if block.children.is_empty() {
+        else if block.get_children().is_empty() {
             self.inadmissible_blocks.push(
                 InadmissibleBlock::new(
-                    block.rows.indices_contained.clone(),
-                    block.columns.indices_contained.clone(),
-                    |i: usize| -> Vec<Cplx> {get_row_or_column(vec![i], block.columns.indices_contained.clone())}
+                    block.get_row_indices().clone(),
+                    block.get_column_indices().clone(),
+                    |i: usize| -> Vec<Cplx> {get_row_or_column(vec![i], block.get_column_indices().clone())}
                 )
             );
         }
         else {
-            for child in block.children {
+            for child in block.get_children() {
                 self.load_from(child, get_row_or_column, tolerance);
             }
         }
@@ -351,42 +202,10 @@ impl HMatrix {
 
 #[cfg(test)]
 mod tests {
-    use std::{collections::HashMap, rc::Rc};
+    use std::collections::HashMap;
     use na::{DMatrix, DVector, Vector3};
     use crate::{preprocess::mesh_data::Node, solve::{h_matrix, tests::generate_random_ab}, Cplx};
-    use super::{Block, Cluster};
 
-    #[test]
-    fn build_cluster_tree() {
-        let mut hmap = HashMap::<usize, usize>::new();
-        let mut nodes = Vec::<Node>::new();
-        let mut i = 0;
-        for j in 0..25 {
-            for k in 0..25 {
-                nodes.push(Node { id: i, coords: Vector3::new(j as f64, k as f64, 0.0)});
-                hmap.insert(i, i);
-                i += 1;
-            }
-        }
-        let tree = Cluster::new_from(&nodes, (0..nodes.len()).collect(), 32, &hmap);
-        approx::assert_relative_eq!(tree.get_diameter(), 33.94, epsilon = 1e-2);
-    }
-    #[test]
-    fn build_block_tree() {
-        let mut hmap = HashMap::<usize, usize>::new();
-        let mut nodes = Vec::<Node>::new();
-        let mut i = 0;
-        for j in 0..25 {
-            for k in 0..25 {
-                nodes.push(Node { id: i, coords: Vector3::new(j as f64, k as f64, 0.0)});
-                hmap.insert(i, i);
-                i += 1;
-            }
-        }
-        let cluster_tree = Rc::new(Cluster::new_from(&nodes, (0..nodes.len()).collect(), 32, &hmap));
-        let block_tree = Block::new_from(cluster_tree.clone(), cluster_tree.clone(), 4.0);
-        assert_eq!(block_tree.children[1].children[1].children[1].admissible, true)
-    }
     fn get_hyperbolic_matrix(n: usize) -> (Vec::<Node>, HashMap::<usize, usize>,
         DMatrix::<Cplx>, DVector::<Cplx>) {
         let mut hmap = HashMap::<usize, usize>::new();
